requirements.txt
boto3==1.9.210
botocore==1.12.210
coverage==4.5.4
docutils==0.15.2
jmespath==0.9.4
numpy==1.17.0
pandas==0.25.1
python-dateutil==2.8.0
pytz==2019.2
s3transfer==0.2.1
six==1.12.0
urllib3==1.25.3
